feat: add verification if pacient has appointments in pacient deletion

+++ src/usecases/pacient/delete-pacient.usecase.ts
@@ -2,9 +2,13 @@ import { PrismaPacientRepository as Repository } from '../../db/prisma-pacient.r
 
 export class DeletePacientUsecase {
   static async execute(cpf: string): Promise<void> {
-    const pacient = await Repository.findByCPF(cpf);
+    const pacient = await Repository.findByCPF(cpf, true);
     if (!pacient) {
       throw new Error('NOT_FOUND');
+    }
+
+    if (pacient.Appointment.length > 0) {
+      throw new Error('PACIENT_HAS_APOINTMENTS');
     }
 
     await Repository.delete(pacient);
